Make an explicit cast

+++ src/main/java/org/osiam/client/oauth/AccessToken.java
@@ -72,7 +72,7 @@ public class AccessToken {
      */
     public boolean isExpired() {
         long now = new Date().getTime();
-        return now > retrievedOn + expiresIn * 1000;
+        return now > retrievedOn + (long) expiresIn * 1000;
     }
 
     /**
